move home screen back to screens folder
